Rearrange test logic
